+++ config.go
@@ -34,14 +34,13 @@ var (
 	queryMethodKey      = attribute.Key("method")
 )
 
-// SpanNameFormatter is an interface that used to format span names.
-// TODO(Sam): change this to function instead of interface.
-type SpanNameFormatter interface {
-	Format(ctx context.Context, method Method, query string) string
-}
+// SpanNameFormatter supports formatting span names.
+type SpanNameFormatter func(ctx context.Context, method Method, query string) string
 
 // AttributesGetter provides additional attributes on spans creation.
 type AttributesGetter func(ctx context.Context, method Method, query string, args []driver.NamedValue) []attribute.KeyValue
+
+type SpanFilter func(ctx context.Context, method Method, query string, args []driver.NamedValue) bool
 
 type config struct {
 	TracerProvider trace.TracerProvider
@@ -116,11 +115,13 @@ type SpanOptions struct {
 
 	// OmitConnectorConnect if set to true will suppress sql.connector.connect spans
 	OmitConnectorConnect bool
+
+	// SpanFilter, if set, will be invoked before each call to create a span. If it returns
+	// false, the span will not be created.
+	SpanFilter SpanFilter
 }
 
-type defaultSpanNameFormatter struct{}
-
-func (f *defaultSpanNameFormatter) Format(_ context.Context, method Method, _ string) string {
+func defaultSpanNameFormatter(_ context.Context, method Method, _ string) string {
 	return string(method)
 }
 
@@ -129,7 +130,7 @@ func newConfig(options ...Option) config {
 	cfg := config{
 		TracerProvider:    otel.GetTracerProvider(),
 		MeterProvider:     otel.GetMeterProvider(),
-		SpanNameFormatter: &defaultSpanNameFormatter{},
+		SpanNameFormatter: defaultSpanNameFormatter,
 	}
 	for _, opt := range options {
 		opt.Apply(&cfg)