+++ CHANGELOG.md
@@ -8,6 +8,40 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+## [0.26.0] - 2023-10-11
+
+### Changed
+
+- Upgrade OTel to version `v1.19.0/v0.42.0`. (#190)
+
+## [0.25.0] - 2023-09-18
+
+### ⚠️ Notice ⚠️
+
+This update contains a breaking change of the type of `SpanNameFormatter`. If you use `SpanNameFormatter` in your code, you need to change the type of `SpanNameFormatter` to function.
+
+The minimum supported Go version is `1.20`.
+
+### Changed
+
+- Upgrade OTel to version `v1.18.0/v0.41.0`. (#184)
+- The type of `SpanNameFormatter` has been changed to function for easier use. (#185)
+
+### Removed
+
+- Support for Go `1.19`. (#186)
+
+## [0.24.0] - 2023-09-08
+
+### Added
+
+- `SpanFilter` configuration in `SpanOptions` to filter spans creation. (#174)
+- Go 1.21 to supported versions. (#180)
+
+### Changed
+
+- Upgrade OTel to version `v1.17.0/v0.40.0`. (#181)
+
 ## [0.23.0] - 2023-05-22
 
 ### Changed
@@ -264,7 +298,10 @@ It contains instrumentation for trace and depends on OTel `v0.18.0`.
 - Example code for a basic usage.
 - Apache-2.0 license.
 
-[Unreleased]: https://github.com/XSAM/otelsql/compare/v0.23.0...HEAD
+[Unreleased]: https://github.com/XSAM/otelsql/compare/v0.26.0...HEAD
+[0.26.0]: https://github.com/XSAM/otelsql/releases/tag/v0.26.0
+[0.25.0]: https://github.com/XSAM/otelsql/releases/tag/v0.25.0
+[0.24.0]: https://github.com/XSAM/otelsql/releases/tag/v0.24.0
 [0.23.0]: https://github.com/XSAM/otelsql/releases/tag/v0.23.0
 [0.22.0]: https://github.com/XSAM/otelsql/releases/tag/v0.22.0
 [0.21.0]: https://github.com/XSAM/otelsql/releases/tag/v0.21.0